fix: resolve all golangci-lint warnings across 20 linters
Enable and fix all issues for: bodyclose, errcheck, errorlint,
exhaustive, gocognit, gocritic, gocyclo, gosec, govet, ineffassign,
intrange, misspell, modernize, nestif, nilerr, nolintlint, perfsprint,
predeclared, staticcheck, unconvert.

Refactored complex functions into helpers and state structs, added
proper error handling for Close/Remove calls, replaced fmt.Sprintf
with faster alternatives, modernized loops and string operations,
added exhaustive switch defaults, and hardened file operations with
filepath.Clean and ReadHeaderTimeout.

Closes gastrolog-3c6g

+++ backend/internal/sysmetrics/sysmetrics.go
@@ -82,6 +82,7 @@ func Memory() MemoryStats {
 
 	rss := peakRSS()
 
+	//nolint:gosec // G115: memory stats are always positive and well within int64 range
 	return MemoryStats{
 		Inuse:        int64(m.HeapInuse + m.StackInuse),
 		RSS:          rss,
@@ -101,7 +102,7 @@ func peakRSS() int64 {
 	if err := syscall.Getrusage(syscall.RUSAGE_SELF, &rusage); err != nil {
 		return 0
 	}
-	rss := int64(rusage.Maxrss)
+	rss := rusage.Maxrss
 	// macOS reports Maxrss in bytes; Linux reports in KB.
 	if runtime.GOOS == "linux" {
 		rss *= 1024
